Database ride injection shell
Created the basic shell of the database injection script. Currently
Queries the database for rides, and checks the rides stored on disk
for discrepancies. Any differences present, kicks off the mechanism
that constructs the queries to add the missing rides. At present, the
script does not add any information to the database. But makes all
necessary queries.